docs(Transition): clarify explorer differences
