+++ docs/app/Examples/modules/Transition/Explorers/TransitionExampleTransitionExplorer.js
@@ -1,24 +1,16 @@
 import React, { Component } from 'react'
 import { Form, Grid, Image, Transition } from 'semantic-ui-react'
 
-const transitions = [
-  'scale',
-  'fade', 'fade up', 'fade down', 'fade left', 'fade right',
-  'horizontal flip', 'vertical flip',
-  'drop',
-  'fly left', 'fly right', 'fly up', 'fly down',
-  'swing left', 'swing right', 'swing up', 'swing down',
-  'browse', 'browse right',
-  'slide down', 'slide up', 'slide right',
-]
+const transitions = ['jiggle', 'flash', 'shake', 'pulse', 'tada', 'bounce']
+
 const options = transitions.map(name => ({ key: name, text: name, value: name }))
 
-export default class TransitionExampleSingleExplorer extends Component {
+export default class TransitionExampleStaticExplorer extends Component {
   state = { animation: transitions[0], duration: 500, visible: true }
 
   handleChange = (e, { name, value }) => this.setState({ [name]: value })
 
-  handleVisibility = () => this.setState({ visible: !this.state.visible })
+  toggleVisibility = () => this.setState({ visible: !this.state.visible })
 
   render() {
     const { animation, duration, visible } = this.state
@@ -43,13 +35,13 @@ export default class TransitionExampleSingleExplorer extends Component {
             type='range'
             value={duration}
           />
-          <Form.Button content={visible ? 'Hide' : 'Show'} onClick={this.handleVisibility} />
+          <Form.Button content='Run' onClick={this.toggleVisibility} />
         </Grid.Column>
 
         <Grid.Column>
-          <Transition.Group animation={animation} duration={duration}>
-            {visible && <Image centered size='small' src='/assets/images/leaves/4.png' />}
-          </Transition.Group>
+          <Transition animation={animation} duration={duration} visible={visible}>
+            <Image centered size='small' src='/assets/images/leaves/5.png' />
+          </Transition>
         </Grid.Column>
       </Grid>
     )
